Move the Floorplan client it the client module.
