+++ __tests__/FloorplanClient.test.js
@@ -1,10 +1,10 @@
-import { FloorplanClient } from '../src';
+import { FloorplanClient } from '../src/client';
 
 
 describe('FloorplanClient', () => {
   it('throws an error if div is not specified in the configuration', () => {
     expect(() => {
       const floorplan = new FloorplanClient({});
-    }).toThrow();
+    }).toThrowError();
   });
 });
